requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
numpy==1.24.3
mediapipe==0.10.8
openai-whisper==20231117
onnxruntime==1.16.3
openai==1.3.7
gunicorn==21.2.0
librosa==0.10.1
soundfile==0.12.1

